tcp: Check if we exceed socket send buffer instead of snd_fack
Summary:
SACK rescue retransmission is supposed to be inserted never
exceeding the right edge of the socket sendbuffer. However,
rarely there appear to be instances, when either the rescue
retransmission is created using an incorrect offset (FIN bit?),
or the send buffer retracts (?). In this case, panic an INVARIANTS
kernel as early as possible to track why this may be happening,
or skip over SACK retransmission processing altogether and revert
to NewReno loss recovery from snd_una.

Reviewers: tuexen, #transport!

Subscribers: imp, melifaro, glebius

Differential Revision: https://reviews.freebsd.org/D35446